backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
beautifulsoup4==4.14.3
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.5
curl_cffi==0.13.0
fastapi==0.129.0
frozendict==2.4.7
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.190.0
google-auth==2.49.0.dev0
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
grpcio==1.78.0
grpcio-status==1.71.2
h11==0.16.0
httplib2==0.31.2
idna==3.11
lxml==6.0.2
multitasking==0.0.12
numpy==2.4.2
packaging==26.0
pandas==3.0.0
patsy==1.0.2
peewee==3.19.0
platformdirs==4.9.2
proto-plus==1.27.1
protobuf==5.29.6
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-multipart==0.0.22
pytz==2025.2
requests==2.32.5
scipy==1.17.0
setuptools==69.5.1
six==1.17.0
soupsieve==2.8.3
starlette==0.52.1
statsmodels==0.14.6
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.41.0
websockets==16.0
yfinance==1.2.0
sqlalchemy>=2.0
psycopg2-binary>=2.9
python-dotenv>=1.0
